`NavigationBar` improvements (#116992)
